Task 10 allure reports UPD
