+++ task_10/tests/test_decorators.py
@@ -1,0 +1,39 @@
+import allure
+from selene.support import by
+from selene.support.shared import browser
+from selene.support.shared.jquery_style import s
+
+
+def test_decorator_steps():
+    open_main_page()
+    search_for_repository("eroshenkoam/allure-example")
+    go_to_repository("eroshenkoam/allure-example")
+    open_issue_tab()
+    should_see_issue_with_name("С Новым Годом")
+
+
+@allure.step("Открываем главную страницу")
+def open_main_page():
+    browser.open("https://github.com")
+
+
+@allure.step("Ищем репозиторий {repo}")
+def search_for_repository(repo):
+    s(".header-search-input").click()
+    s(".header-search-input").send_keys(repo)
+    s(".header-search-input").submit()
+
+
+@allure.step("Переходим по ссылке репозитория {repo}")
+def go_to_repository(repo):
+    s(by.link_text(repo)).click()
+
+
+@allure.step("Открываем таб Issues")
+def open_issue_tab():
+    s("#issues-tab").click()
+
+
+@allure.step("Проверяем наличие Issue с наименованием {name}")
+def should_see_issue_with_name(name):
+    s(by.partial_text(name)).click()
