+++ task_10/tests/test_labels.py
@@ -1,39 +1,25 @@
 import allure
-from selene.support import by
-from selene.support.shared import browser
-from selene.support.shared.jquery_style import s
+from allure_commons.types import Severity
 
 
-def test_decorator_steps():
-    open_main_page()
-    search_for_repository("eroshenkoam/allure-example")
-    go_to_repository("eroshenkoam/allure-example")
-    open_issue_tab()
-    should_see_issue_with_name("С Новым Годом")
+def test_no_labels():
+    pass
 
 
-@allure.step("Открываем главную страницу")
-def open_main_page():
-    browser.open("https://github.com")
+def test_dynamic_labels():
+    allure.dynamic.tag("web")
+    allure.dynamic.severity(Severity.BLOCKER)
+    allure.dynamic.feature("Задачи в репозитории")
+    allure.dynamic.story("Неавторизованный пользователь не может создать задачу в репозитории")
+    allure.dynamic.link("https://github.com", name="Testing")
+    pass
 
 
-@allure.step("Ищем репозиторий {repo}")
-def search_for_repository(repo):
-    s(".header-search-input").click()
-    s(".header-search-input").send_keys(repo)
-    s(".header-search-input").submit()
-
-
-@allure.step("Переходим по ссылке репозитория {repo}")
-def go_to_repository(repo):
-    s(by.link_text(repo)).click()
-
-
-@allure.step("Открываем таб Issues")
-def open_issue_tab():
-    s("#issues-tab").click()
-
-
-@allure.step("Проверяем наличие Issue с наименованием {name}")
-def should_see_issue_with_name(name):
-    s(by.partial_text(name)).click()
+@allure.tag("web")
+@allure.severity(Severity.CRITICAL)
+@allure.label("owner", "eroshenkoam")
+@allure.feature("Задачи в репозитории")
+@allure.story("Авторизованный пользователь может создать задачу в репозитории")
+@allure.link("https://github.com", name="Testing")
+def test_decorator_labels():
+    pass
